pyproject.toml
[tool.poetry]
name = "broh5"
version = "1.0.0"
description = "Browser-based GUI HDF Viewer in Python"
authors = ["Nghia Vo <nvo@bnl.gov>"]
license = "LICENCE"
readme = "README.md"
homepage = "https://github.com/nghia-vo/broh5"
repository = "https://github.com/nghia-vo/broh5"
keywords = ["HDF Viewer", "NXS Viewer", "Browser-based GUI"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
h5py = "*"
pillow = "*"
matplotlib = "*"
nicegui = "^1.3.17"

[tool.poetry.scripts]
broh5 = "broh5.main:main"
